requirements.txt
selenium==4.28.1
pytest==7.0.1
webdriver-manager==4.0.2
faker==37.0.0
allure-pytest